Add PLL lock detection and timeout
Problem: The PLL lock detection is inside a while-true loop. If for
any reason the Spectrum Analyzer fails to achieve a lock then it
will be stuck in an infinite loop.

Solution: A timeout was added to bypass the infinite loop and the
4 most significant bits of the ADC word are set to 0xF as a message
to the user that the PLL failed to achieve lock.

Note: There is also a DEBUG flag that allows you to bypass the lock
detection. That allows the Arduino code to be debugged without the
Spectrum Analyzer board being plugged in.