Rearrange components to remove 3D interference
The shield, Sh1, interferes with the shield on the board above it.
Move Sh1 closer to board edge to eliminate inteference. Plug J1
is open through to the surface of the board so modify the foot-
print to remove ground plane and any traces that might short to
the pins when inserted.