Refactor pwrsupply schematic to show front panel cables & devices
Initially the schematic symbols were being shown on the board for
items that would only be found on the front panel. To properly
show the relationship of the panel mounted parts the symbols were
made to be graphically displayed on the schematic which prevents
them from being added to the PCB. Then 3 connectors were added to
the PCB for ease of connection from the front panel components.
The connectors are conn1 (for the power switch), conn3 for the
power plug, and optionally conn2 (alternate power plug location).