spectrumwvi.pcb: swap C120 and C124 to match schematic
It would have been easier to swap the components on the schematic
without messing around with the board layout. However, the schem-
atic provides a tool 'Autonumber' which can really mess up all
the reference designators if used incorrectly at a later date.
Moving the components on the PCB instead is slightly safer.