Remove CommandBits: masking value not needed
serialWordAsBytes is a byte pointer to the 32 bit serialWord.
The second byte, serialWordAsBytes[1], contains the upper 5 bits
as the Command and the lower 3 bits as the Address. To get the
Command-Value we are masking off the 3 lower bits and then right
shifting 3 bits. Doing a right-shift makes the masking redundant