Rename CmdProcessor.ino to CmdProcessor.ini.cpp
In another project someone submitted a request to move the project
over to platformio. Once I moved this over to platformio I had to
rename the sketch to work in the Arduino IDE again and vice-versa.

It seems that if you rename the file to mySketch.ino.cpp then you
can use it in either dev tool without renaming. You'll still need
the guard clause:
+----------------------
| #ifdef PLATFORMIO
| #include <Arduino.h>
| #endif
+----------------------