Fix CmdProcessor.cpp sending too many data packets to PC application
CmdProcessor should be handling *every* single Serial.write. There
were 2 Serial.write() blocks in SpecAnn.cpp and this takes care of
one of them. The problem was that the writes were at the top level
of a while-loop and needed to guard against writing and rewriting
old data to the serial port.